Introducing participant model factory pattern
