+++ src/main/scala/edu/ie3/simona/model/participant2/ParticipantModel.scala
@@ -6,6 +6,7 @@
 
 package edu.ie3.simona.model.participant2
 
+import edu.ie3.datamodel.models.input.system.SystemParticipantInput
 import edu.ie3.datamodel.models.result.system.SystemParticipantResult
 import edu.ie3.simona.agent.participant.data.Data.PrimaryData.{
   ComplexPower,
@@ -18,6 +19,7 @@ import edu.ie3.simona.model.participant2.ParticipantModel.{
 }
 import edu.ie3.simona.agent.participant2.ParticipantAgent
 import edu.ie3.simona.agent.participant2.ParticipantAgent.ParticipantRequest
+import edu.ie3.simona.config.RuntimeConfig.BaseRuntimeConfig
 import edu.ie3.simona.model.participant.control.QControl
 import edu.ie3.simona.service.ServiceType
 import edu.ie3.util.scala.quantities.DefaultQuantities.zeroKW
@@ -212,6 +214,13 @@ abstract class ParticipantModel[
 }
 
 object ParticipantModel {
+
+  trait ParticipantModelFactory[
+      S <: SystemParticipantInput,
+      C <: BaseRuntimeConfig,
+  ] {
+    def create(input: S, config: C): ParticipantModel[_, _]
+  }
 
   trait OperatingPoint {
