+++ src/main/scala/edu/ie3/simona/model/participant2/load/profile/ProfileLoadModel.scala
@@ -59,7 +59,7 @@ class ProfileLoadModel(
 
 object ProfileLoadModel {
 
-  def apply(input: LoadInput, config: LoadRuntimeConfig): ProfileLoadModel = {
+  def create(input: LoadInput, config: LoadRuntimeConfig): ProfileLoadModel = {
 
     val loadProfileStore = LoadProfileStore()
