+++ src/test/scala/edu/ie3/simona/model/participant2/StorageModelSpec.scala
@@ -79,7 +79,7 @@ class StorageModelSpec extends UnitSpec with Matchers {
   def buildStorageModel(
       targetSoc: Option[Double] = Option.empty
   ): StorageModel =
-    StorageModel.apply(
+    StorageModel.create(
       inputModel,
       StorageRuntimeConfig(targetSoc = targetSoc),
     )
