+++ src/test/scala/edu/ie3/simona/model/participant2/load/FixedLoadModelSpec.scala
@@ -33,7 +33,7 @@ class FixedLoadModelSpec extends UnitSpec with LoadInputTestData {
           reference = reference,
         )
 
-        val model = FixedLoadModel(
+        val model = FixedLoadModel.create(
           loadInput,
           config,
         )
